no s2e

+++ reporte/src/main/java/Reporte/reporte/model/ReporteRequest.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 
 public class ReporteRequest {
 
-    private Integer usuarioId;
+    private long usuarioId;
     private String descripcion;
 
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
@@ -22,7 +22,7 @@ public class ReporteRequest {
         return usuarioId;
     }
 
-    public void setUsuarioId(Integer usuarioId) {
+    public void setUsuarioId(long usuarioId) {
         this.usuarioId = usuarioId;
     }
